requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
crawl4ai
openai
pdfplumber
python-docx
aiofiles
python-multipart
newspaper3k
lxml[html_clean]
